# backend/openapi/healthv1.yml
# yaml-language-server: $schema=https://raw.githubusercontent.com/OAI/OpenAPI-Specification/ba055ca00cbd735dceef4ed7c9db024cd9bfcd1a/schemas/v3.1/schema.json
openapi: 3.1.0
info:
  title: Health API
  version: 1.0.0
  servers:
    - url: https://libdb.so/api/health/v1

paths:
  /current.json:
    summary: Get the pet's current health data
    description: |
      Get the pet's current health data.
    get:
      operationId: getCurrentHealth
      responses:
        "200":
          description: |
            Pet's current health data in JSON format.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentHealth"

  /history.json:
    summary: Get the pet's health history
    description: |
      Get the pet's health history for a given time range.
    get:
      operationId: getHealthHistory
      parameters:
        - name: form
          description: |
            The start of the time range to fetch data from.

            The server may reject `from` parameters that are too far in the past.
          in: query
          schema:
            $ref: "#/components/schemas/Timestamp"
        - name: to
          description: |
            The end of the time range to fetch data from. This timestamp must be
            after `from`. 

            The server may reject `to` parameters and only return data up to the
            present time.
          in: query
          schema:
            $ref: "#/components/schemas/Timestamp"
      responses:
        "200":
          description: |
            Pet's health history data in JSON format.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthHistory"
        "501":
          description: |
            The server does not support this endpoint.

components:
  schemas:
    CurrentHealth:
      description: |
        Current health data of the pet.
      type: object
      properties:
        heartRate:
          description: Current heart rate of the pet
          allOf:
            - $ref: "#/components/schemas/HeartRate"
        steps:
          description: Current steps taken by the pet
          allOf:
            - $ref: "#/components/schemas/StepsTaken"
        sleep:
          description: Current sleep status of the pet
          allOf:
            - $ref: "#/components/schemas/SleepStatus"
        estrogen:
          description: Current estrogen level of the pet
          allOf:
            - $ref: "#/components/schemas/EstrogenLevel"

    HealthHistory:
      description: |
        History of the pet's health data.

        This data is returned from /history.json and may cover the time range
        given in the parameters.
      type: object
      properties:
        heartRate:
          description: Heart rate of the pet
          allOf:
            - $ref: "#/components/schemas/HeartRateHistory"
        steps:
          description: Steps taken by the pet
          allOf:
            - $ref: "#/components/schemas/StepsHistory"
        sleep:
          description: Sleep data of the pet
          allOf:
            - $ref: "#/components/schemas/SleepHistory"
        estrogen:
          description: Estrogen level of the pet
          allOf:
            - $ref: "#/components/schemas/EstrogenHistory"

    ##
    ## Sleep Data
    ##

    SleepHistory:
      description: |
        Sleep history of the pet.

        The sleep history is represented as a list of sleep or awake periods,
        optionally labeled with the sleep type.
      type: array
      items:
        allOf:
          - type: object
            required: [from, to, asleep]
            properties:
              from:
                description: Start of the sleep period
                allOf:
                  - $ref: "#/components/schemas/Timestamp"
              to:
                description: End of the sleep period
                allOf:
                  - $ref: "#/components/schemas/Timestamp"
          - $ref: "#/components/schemas/SleepStatus"

    SleepStatus:
      description: |
        The current sleep status of the pet.

        The sleep status is represented as a single sleep or awake period,
        optionally labeled with the sleep type. It is meant to be used to
        describe the current sleep state of the pet.
      type: object
      required: [asleep]
      properties:
        asleep:
          description: Whether the pet is asleep
          type: boolean
        sleepType:
          description: Type of sleep
          allOf:
            - $ref: "#/components/schemas/SleepType"

    SleepType:
      description: Type of sleep
      type: string
      enum:
        - awake
        - rem
        - light
        - deep

    ##
    ## Estrogen Data
    ##

    EstrogenHistory:
      description: |
        Estimated estrogen levels.

        The estrogen levels are represented as a list of data points describing
        the estrogen levels in ng/dL at different timestamps. The timestamps may
        not be equally spaced.
      type: object
      required: [type]
      properties:
        type:
          type: string
          description: |
            Type of estrogen (e.g. patches, pills, injection).
        levels:
          description: |
            Estimated estrogen levels of the pet over time in ng/dL.
          type: array
          items:
            $ref: "#/components/schemas/DataPoint"
        dosages:
          type: array
          description: |
            When the estrogen was administered, represented as a list of
            timestamps.
          items:
            $ref: "#/components/schemas/Timestamp"

    EstrogenLevel:
      description: |
        Current estimated estrogen level in ng/dL
      type: number
      format: double

    ##
    ## Steps Data
    ##

    StepsHistory:
      description: |
        Steps taken by the pet over time.
      type: object
      required: [data]
      properties:
        data:
          description: |
            Data points of the number of steps taken at different timestamps.
          type: array
          items:
            $ref: "#/components/schemas/DataPoint"

    StepsTaken:
      description: |
        Number of steps taken by the pet over a period of time.
      type: object
      properties:
        from:
          description: Start of the period
          allOf:
            - $ref: "#/components/schemas/Timestamp"
        to:
          description: End of the period
          allOf:
            - $ref: "#/components/schemas/Timestamp"
        count:
          description: Number of steps taken
          type: integer
          format: int32

    ##
    ## Heart Rate Data
    ##

    HeartRateHistory:
      description: |
        Heart rate of the pet over time.
      type: object
      required: [data]
      properties:
        data:
          description: |
            Data points of the heart rate at different timestamps.

            Each data point is the average heart rate in beats per minute over a
            period of time, usually the delta between a timestamp and its previous
            one.
          type: array
          items:
            $ref: "#/components/schemas/DataPoint"

    HeartRate:
      description: |
        Current heart rate of the pet in beats per minute.
      type: number
      format: double

    ##
    ## Common Types
    ##

    DataPoint:
      type: object
      required:
        - timestamp
        - value
      properties:
        timestamp:
          $ref: "#/components/schemas/Timestamp"
        value:
          type: number
          format: double
          description: Value of the data point

    Timestamp:
      type: string
      format: date-time