# backend/openapi/healthv1api/openapi.gen.go
// Package healthv1api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version 2.3.0 DO NOT EDIT.
package healthv1api

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	strictnethttp "github.com/oapi-codegen/runtime/strictmiddleware/nethttp"
)

// Defines values for SleepType.
const (
	Awake SleepType = "awake"
	Deep  SleepType = "deep"
	Light SleepType = "light"
	Rem   SleepType = "rem"
)

// CurrentHealth Current health data of the pet.
type CurrentHealth struct {
	// Estrogen Current estrogen level of the pet
	Estrogen *EstrogenLevel `json:"estrogen,omitempty"`

	// HeartRate Current heart rate of the pet
	HeartRate *HeartRate `json:"heartRate,omitempty"`

	// Sleep Current sleep status of the pet
	Sleep *SleepStatus `json:"sleep,omitempty"`

	// Steps Current steps taken by the pet
	Steps *StepsTaken `json:"steps,omitempty"`
}

// DataPoint defines model for DataPoint.
type DataPoint struct {
	Timestamp Timestamp `json:"timestamp"`

	// Value Value of the data point
	Value float64 `json:"value"`
}

// EstrogenHistory Estimated estrogen levels.
//
// The estrogen levels are represented as a list of data points describing
// the estrogen levels in ng/dL at different timestamps. The timestamps may
// not be equally spaced.
type EstrogenHistory struct {
	// Dosages When the estrogen was administered, represented as a list of
	// timestamps.
	Dosages *[]Timestamp `json:"dosages,omitempty"`

	// Levels Estimated estrogen levels of the pet over time in ng/dL.
	Levels *[]DataPoint `json:"levels,omitempty"`

	// Type Type of estrogen (e.g. patches, pills, injection).
	Type string `json:"type"`
}

// EstrogenLevel Current estimated estrogen level in ng/dL
type EstrogenLevel = float64

// HealthHistory History of the pet's health data.
//
// This data is returned from /history.json and may cover the time range
// given in the parameters.
type HealthHistory struct {
	// Estrogen Estrogen level of the pet
	Estrogen *EstrogenHistory `json:"estrogen,omitempty"`

	// HeartRate Heart rate of the pet
	HeartRate *HeartRateHistory `json:"heartRate,omitempty"`

	// Sleep Sleep data of the pet
	Sleep *SleepHistory `json:"sleep,omitempty"`

	// Steps Steps taken by the pet
	Steps *StepsHistory `json:"steps,omitempty"`
}

// HeartRate Current heart rate of the pet in beats per minute.
type HeartRate = float64

// HeartRateHistory Heart rate of the pet over time.
type HeartRateHistory struct {
	// Data Data points of the heart rate at different timestamps.
	//
	// Each data point is the average heart rate in beats per minute over a
	// period of time, usually the delta between a timestamp and its previous
	// one.
	Data []DataPoint `json:"data"`
}

// SleepHistory Sleep history of the pet.
//
// The sleep history is represented as a list of sleep or awake periods,
// optionally labeled with the sleep type.
type SleepHistory = []struct {
	// Asleep Whether the pet is asleep
	Asleep bool `json:"asleep"`

	// From Start of the sleep period
	From Timestamp `json:"from"`

	// SleepType Type of sleep
	SleepType *SleepType `json:"sleepType,omitempty"`

	// To End of the sleep period
	To Timestamp `json:"to"`
}

// SleepStatus The current sleep status of the pet.
//
// The sleep status is represented as a single sleep or awake period,
// optionally labeled with the sleep type. It is meant to be used to
// describe the current sleep state of the pet.
type SleepStatus struct {
	// Asleep Whether the pet is asleep
	Asleep bool `json:"asleep"`

	// SleepType Type of sleep
	SleepType *SleepType `json:"sleepType,omitempty"`
}

// SleepType Type of sleep
type SleepType string

// StepsHistory Steps taken by the pet over time.
type StepsHistory struct {
	// Data Data points of the number of steps taken at different timestamps.
	Data []DataPoint `json:"data"`
}

// StepsTaken Number of steps taken by the pet over a period of time.
type StepsTaken struct {
	// Count Number of steps taken
	Count *int32 `json:"count,omitempty"`

	// From Start of the period
	From *Timestamp `json:"from,omitempty"`

	// To End of the period
	To *Timestamp `json:"to,omitempty"`
}

// Timestamp defines model for Timestamp.
type Timestamp = time.Time

// GetHealthHistoryParams defines parameters for GetHealthHistory.
type GetHealthHistoryParams struct {
	// Form The start of the time range to fetch data from.
	//
	// The server may reject `from` parameters that are too far in the past.
	Form *Timestamp `form:"form,omitempty" json:"form,omitempty"`

	// To The end of the time range to fetch data from. This timestamp must be
	// after `from`.
	//
	// The server may reject `to` parameters and only return data up to the
	// present time.
	To *Timestamp `form:"to,omitempty" json:"to,omitempty"`
}

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (GET /current.json)
	GetCurrentHealth(w http.ResponseWriter, r *http.Request)

	// (GET /history.json)
	GetHealthHistory(w http.ResponseWriter, r *http.Request, params GetHealthHistoryParams)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// (GET /current.json)
func (_ Unimplemented) GetCurrentHealth(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /history.json)
func (_ Unimplemented) GetHealthHistory(w http.ResponseWriter, r *http.Request, params GetHealthHistoryParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetCurrentHealth operation middleware
func (siw *ServerInterfaceWrapper) GetCurrentHealth(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetCurrentHealth(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// GetHealthHistory operation middleware
func (siw *ServerInterfaceWrapper) GetHealthHistory(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetHealthHistoryParams

	// ------------- Optional query parameter "form" -------------

	err = runtime.BindQueryParameter("form", true, false, "form", r.URL.Query(), &params.Form)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "form", Err: err})
		return
	}

	// ------------- Optional query parameter "to" -------------

	err = runtime.BindQueryParameter("form", true, false, "to", r.URL.Query(), &params.To)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "to", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHealthHistory(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/current.json", wrapper.GetCurrentHealth)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/history.json", wrapper.GetHealthHistory)
	})

	return r
}

type GetCurrentHealthRequestObject struct {
}

type GetCurrentHealthResponseObject interface {
	VisitGetCurrentHealthResponse(w http.ResponseWriter) error
}

type GetCurrentHealth200JSONResponse CurrentHealth

func (response GetCurrentHealth200JSONResponse) VisitGetCurrentHealthResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetHealthHistoryRequestObject struct {
	Params GetHealthHistoryParams
}

type GetHealthHistoryResponseObject interface {
	VisitGetHealthHistoryResponse(w http.ResponseWriter) error
}

type GetHealthHistory200JSONResponse HealthHistory

func (response GetHealthHistory200JSONResponse) VisitGetHealthHistoryResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetHealthHistory501Response struct {
}

func (response GetHealthHistory501Response) VisitGetHealthHistoryResponse(w http.ResponseWriter) error {
	w.WriteHeader(501)
	return nil
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {

	// (GET /current.json)
	GetCurrentHealth(ctx context.Context, request GetCurrentHealthRequestObject) (GetCurrentHealthResponseObject, error)

	// (GET /history.json)
	GetHealthHistory(ctx context.Context, request GetHealthHistoryRequestObject) (GetHealthHistoryResponseObject, error)
}

type StrictHandlerFunc = strictnethttp.StrictHTTPHandlerFunc
type StrictMiddlewareFunc = strictnethttp.StrictHTTPMiddlewareFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// GetCurrentHealth operation middleware
func (sh *strictHandler) GetCurrentHealth(w http.ResponseWriter, r *http.Request) {
	var request GetCurrentHealthRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetCurrentHealth(ctx, request.(GetCurrentHealthRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetCurrentHealth")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetCurrentHealthResponseObject); ok {
		if err := validResponse.VisitGetCurrentHealthResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetHealthHistory operation middleware
func (sh *strictHandler) GetHealthHistory(w http.ResponseWriter, r *http.Request, params GetHealthHistoryParams) {
	var request GetHealthHistoryRequestObject

	request.Params = params

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetHealthHistory(ctx, request.(GetHealthHistoryRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetHealthHistory")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetHealthHistoryResponseObject); ok {
		if err := validResponse.VisitGetHealthHistoryResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}